Combined telemetry extensions and made fields readonly.
